+++ include/roq/tool.hpp
@@ -6,16 +6,16 @@
 #include <string_view>
 #include <vector>
 
-#include "roq/logging.h"
+#include "roq/logging.hpp"
 
 namespace roq {
 
 //! Convenience class to wrap \ref roq::Logger, absl::flags, etc.
-class ROQ_PUBLIC Service {
+class ROQ_PUBLIC Tool {
  public:
   struct Info final {
     std::string_view description;
-    std::string_view package_name = {};
+    std::string_view package_name;
     std::string_view build_version = {};
     std::string_view build_number = {};
     std::string_view build_type = {};
@@ -24,12 +24,12 @@ class ROQ_PUBLIC Service {
     std::string_view compile_time = __TIME__;
   };
 
-  Service(int argc, char **argv, const Info &);
+  Tool(int argc, char **argv, const Info &);
 
-  Service(const Service &) = delete;
-  Service(Service &&) = default;
+  Tool(const Tool &) = delete;
+  Tool(Tool &&) = default;
 
-  virtual ~Service();
+  virtual ~Tool();
 
   //! The main function
   int run();
@@ -39,9 +39,6 @@ class ROQ_PUBLIC Service {
 
  private:
   std::vector<char *> args_;
-  std::string package_name_;
-  std::string build_version_;
-  std::string build_number_;
   std::string build_type_;
   std::string git_hash_;
   std::string compile_date_;