+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Changed
+
+* Re-introduce Abseil Flags path, but exclude build prefix
+
 ## 0.7.9 &ndash; 2021-12-08
 
 ## 0.7.8 &ndash; 2021-11-02