+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Changed                                                                                                                
+
+* Conda packaging should pin versions to match x.x.x (#189)
+
 ## 0.8.2 &ndash; 2022-02-18
 
 ### Changed